Refactored project file structure to avoid circular imports; (#4220)
