+++ lib/adapters/http.js
@@ -13,7 +13,7 @@ var zlib = require('zlib');
 var VERSION = require('./../env/data').version;
 var createError = require('../core/createError');
 var enhanceError = require('../core/enhanceError');
-var defaults = require('../defaults');
+var transitionalDefaults = require('../defaults/transitional');
 var Cancel = require('../cancel/Cancel');
 
 var isHttps = /https:?/;
@@ -365,7 +365,7 @@ module.exports = function httpAdapter(config) {
         } else {
           timeoutErrorMessage = 'timeout of ' + config.timeout + 'ms exceeded';
         }
-        var transitional = config.transitional || defaults.transitional;
+        var transitional = config.transitional || transitionalDefaults;
         reject(createError(
           timeoutErrorMessage,
           config,
