+++ lib/core/transformData.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var utils = require('./../utils');
-var defaults = require('./../defaults');
+var defaults = require('../defaults');
 
 /**
  * Transform the data for a request or a response
