+++ lib/defaults/transitional.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  silentJSONParsing: true,
+  forcedJSONParsing: true,
+  clarifyTimeoutError: false
+};
